api branchpromotion
api branchpromotion

+++ BackEnd/park_management/src/main/java/park/management/com/vn/entity/BranchPromotion.java
@@ -35,7 +35,6 @@ public class BranchPromotion extends BaseEntity {
   private LocalDateTime to;
 
   @Column(name = "is_active")
-  private Integer isActive;
-
+  private Long isActive;
 
 }
